+++ trunk/openvibe-applications/designer/trunc/src/ovdCBoxConfigurationDialog.cpp
@@ -180,7 +180,7 @@ static void on_button_load_clicked(::GtkButton* pButton, gpointer pUserData)
 			size_t l_iBufferLen=0;
 			size_t l_iFileLen;
 			l_oFile.seekg(0, ios::end);
-			l_iFileLen=l_oFile.tellg();
+			l_iFileLen=(size_t)l_oFile.tellg();
 			l_oFile.seekg(0, ios::beg);
 			while(l_iFileLen && l_bStatusOk)
 			{
